Add `test_cuda_multigpu`
Mostly refactor, that moves all the tests that benefit from multigpu
environment into its own file
